chore(widget-builder): Remove old code
