--- static/app/views/dashboards/widgetBuilder/buildSteps/dataSetStep.spec.tsx
@@ -1,26 +1,0 @@
-import {OrganizationFixture} from 'sentry-fixture/organization';
-
-import {render, screen} from 'sentry-test/reactTestingLibrary';
-
-import {DisplayType} from 'sentry/views/dashboards/types';
-import {DataSetStep} from 'sentry/views/dashboards/widgetBuilder/buildSteps/dataSetStep';
-import {DataSet} from 'sentry/views/dashboards/widgetBuilder/utils';
-
-describe('DataSetStep', () => {
-  it('renders the spans dataset with the EAP feature flag', () => {
-    render(
-      <DataSetStep
-        dataSet={DataSet.ISSUES}
-        displayType={DisplayType.TABLE}
-        onChange={jest.fn()}
-      />,
-      {
-        organization: OrganizationFixture({
-          features: ['visibility-explore-view'],
-        }),
-      }
-    );
-
-    expect(screen.getByText('Spans')).toBeInTheDocument();
-  });
-});
