--- static/app/views/dashboards/widgetBuilder/buildSteps/yAxisStep/yAxisSelector/addButton.tsx
@@ -1,17 +1,0 @@
-import type {MouseEventHandler} from 'react';
-
-import {Button} from 'sentry/components/core/button';
-import {IconAdd} from 'sentry/icons';
-
-interface Props {
-  onAdd: MouseEventHandler<Element>;
-  title: string;
-}
-
-export function AddButton({title, onAdd}: Props) {
-  return (
-    <Button size="sm" onClick={onAdd} icon={<IconAdd isCircled />}>
-      {title}
-    </Button>
-  );
-}
