--- static/app/views/dashboards/widgetBuilder/buildSteps/yAxisStep/yAxisSelector/deleteButton.tsx
@@ -1,22 +1,0 @@
-import type {MouseEventHandler} from 'react';
-
-import {Button} from 'sentry/components/core/button';
-import {IconDelete} from 'sentry/icons';
-import {t} from 'sentry/locale';
-
-interface Props {
-  onDelete: MouseEventHandler<Element>;
-}
-
-export function DeleteButton({onDelete}: Props) {
-  return (
-    <Button
-      size="zero"
-      borderless
-      onClick={onDelete}
-      icon={<IconDelete />}
-      title={t('Remove this Y-Axis')}
-      aria-label={t('Remove this Y-Axis')}
-    />
-  );
-}
