+++ static/app/views/dashboards/widgets/bigNumberWidget/thresholdsIndicator.tsx
@@ -3,8 +3,8 @@ import styled from '@emotion/styled';
 
 import type {Polarity} from 'sentry/components/percentChange';
 import {normalizeUnit} from 'sentry/views/dashboards/utils';
+import type {ThresholdsConfig} from 'sentry/views/dashboards/widgetBuilder/buildSteps/thresholdsStep/thresholds';
 import {ThresholdsHoverWrapper} from 'sentry/views/dashboards/widgetBuilder/buildSteps/thresholdsStep/thresholdsHoverWrapper';
-import type {ThresholdsConfig} from 'sentry/views/dashboards/widgetBuilder/buildSteps/thresholdsStep/thresholdsStep';
 
 type ValidThresholds = {
   max_values: {
